% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/int2base36.R
\name{int2base36}
\alias{int2base36}
\title{Converts an integer to a base36 string}
\usage{
int2base36(int)
}
\arguments{
\item{int}{Integer input}
}
\description{
Converts an integer to a base36 string
}
\details{
Note that int \emph{must} be an integer. You can convert a
numeric object to integer with \code{as.integer}.

Note this function is not vectorized. Wrap it in sapply if needed.
Thanks to Joshua Ulrich for providing the source:
https://stackoverflow.com/questions/36036404/convert-an-integer-to-base36
}
