% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msq2acres.R
\name{msq2acres}
\alias{msq2acres}
\title{Meters squared to acres}
\usage{
msq2acres(x)
}
\arguments{
\item{x}{A value in feet to convert to meters-squared}
}
\value{
The value in acres
}
\description{
Converts meters squared to acres
}
