% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{na2val}
\alias{na2val}
\title{Convert NA values to a number}
\usage{
na2val(x, val = 0)
}
\arguments{
\item{x}{A value to be tested}

\item{val}{The value to replace NAs with}
}
\value{
The value
}
\description{
Convert NA values to a number
}
\details{
This is an internal function used to convert NA to something else. This
is useful in if() statements when checking the value of an object which may possibly not exist.

This function is vectorized.
}
\keyword{internal}
