% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/null2na.R
\name{null2na}
\alias{null2na}
\title{Convert NULL values to NA}
\usage{
null2na(x)
}
\arguments{
\item{x}{A value}
}
\value{
The value
}
\description{
Convert NULL values to NA
}
\details{
This is an internal function used to convert NULL values to NA. This
is useful in if() statements when checking the value of an object which may possibly not exist.

This function is *not* vectorized.
}
