% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uas_grp_generics.R
\name{print.uas_grp}
\alias{print.uas_grp}
\title{Prints a UAS Image Collection Groups object}
\usage{
\method{print}{uas_grp}(x, ...)
}
\arguments{
\item{x}{A UAS Image Collection Groups object}

\item{...}{unused}
}
\description{
Prints a UAS Image Collection Groups object
}
