% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.uas_info.R
\name{print.uas_info}
\alias{print.uas_info}
\title{Print a summary of a UAS Image Collection object}
\usage{
\method{print}{uas_info}(x, meta_extra = TRUE, ...)
}
\arguments{
\item{x}{A UAS image collection object}

\item{meta_extra}{Show additional meta data info}

\item{...}{Other arguments (unused)}
}
\description{
Print a summary of a UAS Image Colletion object
}
\details{
Prints a summary of a drone images metadata collection object
}
\seealso{
\link{uas_info}
}
