% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.uas_info.R
\name{summary.uas_info}
\alias{summary.uas_info}
\title{Show a summary of a drone images metadata collection object}
\usage{
\method{summary}{uas_info}(object, ...)
}
\arguments{
\item{object}{A drone images metadata object}

\item{...}{Other arguments (unused)}
}
\description{
Prints a summary of a drone images metadata collection object
}
\details{
Prints a summary of a drone images metadata collection object
}
\seealso{
\link{uas_info}
}
