% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uas_dirs_find.R
\name{uas_dirs_find}
\alias{uas_dirs_find}
\title{Search directories for images}
\usage{
uas_dirs_find(x, ext = c("jpg", "tif"), min_images = 3, exclude_tb = TRUE)
}
\arguments{
\item{x}{The base directory}

\item{ext}{A character vector of the file extensions to search for (excluding the '.')}

\item{min_images}{The minimum number of image files a directory must contain to be included in the results}

\item{exclude_tb}{Exclude folders called 'tb' (thumbnails)}
}
\value{
A tibble with 3 columns: \code{path}, \code{ext}, and \code{num_files}.
}
\description{
Find sub-directories which contain images
}
\details{
This function recurses through sub-directories of `x` and returns those that contain image files.
This can be used to help identify folders that need to be cataloged.

Only directories that have at least \code{min_images} image files will be returned.
If \code{exclude_tb = TRUE}, directories named `tb` (which is where \code{\link{uas_thumbnails_make}} saves thumbnail
images) are excluded. Hidden directories and directories that start with `.` are excluded.
}
