% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uas_exp.R
\name{uas_exp}
\alias{uas_exp}
\title{Export spatial data from a UAS image collection}
\usage{
uas_exp(
  x,
  ctr = FALSE,
  fp = FALSE,
  mcp = FALSE,
  shp_dir = NULL,
  create_dir = TRUE,
  overwrite = FALSE,
  quiet = FALSE
)
}
\arguments{
\item{x}{A list of class 'uas_info'}

\item{ctr}{Export the image centroids as a Shapefile, T/F or a filename}

\item{fp}{Export the image footprints as a Shapefile, T/F or a filename}

\item{mcp}{Export the minimum convex polygon of the image footprints as a Shapefile, T/F or a filename}

\item{shp_dir}{The directory where the Shapefiles should be saved. If NULL, they will be saved in a 'map' sub-directory of the image folder}

\item{create_dir}{Create the output directory if it doesn't exist, T/F}

\item{overwrite}{Overwrite existing files, T/F}

\item{quiet}{Supress messages and printing of the pandoc command line, T/F}
}
\value{
A vector of Shapefile filenames
}
\description{
Export spatial data from a UAS image collection
}
\details{
\code{ctr}, \code{fp}, and \code{mcp} can be logical (TRUE/FALSE) or a filename
(minus the shp extension). If logical values are passed, a Shapefile name will
be constructed based on the name of parent folder. You can specify which directory
the Shapefiles will be exported to using the \code{shp_dir} argument. The default
is to save them in a sub-directory of the images directory called 'map'.
}
\seealso{
\link{uas_info}
}
