% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uas_exp_kml.R, R/uas_exp_shp.R
\name{uas_exp_kml}
\alias{uas_exp_kml}
\alias{uas_exp_shp}
\title{Export flight area and camera locations for GIS}
\usage{
uas_exp_kml(
  x,
  flt = NULL,
  ctr = FALSE,
  fp = FALSE,
  mcp = FALSE,
  combine_feats = FALSE,
  combine_fn = NULL,
  output_dir = NULL,
  out_fnbase = NULL,
  create_dir = TRUE,
  overwrite = FALSE,
  quiet = FALSE,
  flt_idx = deprecated()
)

uas_exp_shp(
  x,
  flt = NULL,
  ctr = FALSE,
  fp = FALSE,
  mcp = FALSE,
  combine_feats = FALSE,
  combine_fn = NULL,
  output_dir = NULL,
  out_fnbase = NULL,
  create_dir = TRUE,
  overwrite = FALSE,
  quiet = FALSE,
  flt_idx = deprecated()
)
}
\arguments{
\item{x}{A list of class 'uas_info'}

\item{flt}{Flight(s) in x to process (character or numeric vector, default is all)}

\item{ctr}{Export the image centroids, Logical}

\item{fp}{Export the image footprints, Logical}

\item{mcp}{Export the minimum convex polygon of the image footprints, logical}

\item{combine_feats}{Combine features into a single Shapefile / KML, logical}

\item{combine_fn}{File name (minus path and extension) for the combined features layer}

\item{output_dir}{The output directory. If NULL, the layers will be saved in a 'map' sub-directory of the image folder}

\item{out_fnbase}{The base of output filenames}

\item{create_dir}{Create the output directory if it doesn't exist, logical}

\item{overwrite}{Overwrite existing files, Logical}

\item{quiet}{Suppress messages, Logical}

\item{flt_idx}{`r lifecycle::badge("deprecated")` Use `flt` instead}
}
\value{
A vector of filenames
}
\description{
Export geometry(s) from a flight to KML and Shapefile
}
\details{
\code{flt} allows you to specify a subset of image folders in \code{x} to process. You can pass a vector of flight names (use names(x)
to see what those are) or integers.

\code{ctr}, \code{fp}, and \code{mcp} (all TRUE/FALSE) specify which geometry(s) to export.
Output file names will be generated from the flight metadata saved in the uas_info object. Alternately,
you can pass the base of a file name using \code{out_fnbase}. You can specify the output directory
with \code{output_dir}. The default is to save them in a sub-directory of the images directory
called 'map', which will be created if needed.

If \code{combine_feats = TRUE}, the geometries from the folders in x will be combined into a single KML or Shapefile.
For this to happen, you must also pass a value for \code{combine_fn} (a base for the file name of the combined features).
}
\section{Functions}{
\itemize{
\item \code{uas_exp_shp()}: Export flight info to Shapefile

}}
\seealso{
\code{\link{uas_info}}, \code{\link{uas_report}}
}
