% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uas_cache.R
\name{uas_getcache}
\alias{uas_getcache}
\alias{uas_setcache}
\title{Manage cache directory}
\usage{
uas_getcache(default = TRUE, quiet = FALSE)

uas_setcache(dir = "~/.R/uasimg", write = FALSE, quiet = FALSE)
}
\arguments{
\item{default}{Use a default cache directory if no other has been set}

\item{quiet}{Suppress messages}

\item{dir}{The directory for  cached EXIF data (must exist)}

\item{write}{Write directory location to .Renviron}
}
\description{
View and set the directory where extracted EXIF data are cached
}
\details{
Extracting exif data from a large number of images can take awhile. To avoid having
to do this more than once, the results can be saved or cached to a directory of your choice.
The next time you call \code{\link{uas_info}}, R will first look to see if EXIF data for that
image collection has already been generated, and if so use it instead of running
exiftool again.

Cached results are saved as native R objects. File names encode the name of the
image directory as well as the total number of images and file size. Hence if images are
removed or added from a directory, any cached results will be nullified and exiftool will
run again. The cached data does not include supplemental metadata, such as the
collection name or data URI.

If \code{default = TRUE}, a default directory for the cache (\emph{~/.R/uasimg}) will be used
if another one has not already been set.
}
\section{Functions}{
\itemize{
\item \code{uas_setcache}: Set cache directory
}}

\seealso{
\code{\link{uas_info}}
}
