% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uas_grp_flt.R
\name{uas_grp_flt}
\alias{uas_grp_flt}
\title{Parse an image collection into individual flights}
\usage{
uas_grp_flt(
  x,
  interflt_val = 10,
  interflt_units = c("med_int", "secs")[1],
  init_fltnum = 1,
  min_images = 5,
  cross_dirs = TRUE,
  options = NULL,
  quiet = FALSE
)
}
\arguments{
\item{x}{A list of class 'uas_info'}

\item{interflt_val}{Time value between flights}

\item{interflt_units}{Time units between flights, see details "med_int" or "secs"}

\item{init_fltnum}{Initial flight number}

\item{min_images}{Minimum number of images to be considered a flight}

\item{cross_dirs}{Parse the images in directories collectively}

\item{options}{Options and overrides for creating groups}

\item{quiet}{Suppress messages}
}
\description{
Parse an image collection into individual flights
}
\details{
cross_dirs means images in all directories will be examined when look for flights. This
would be appropriate if the images from one flight were spread across multiple folders.
}
