% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uas_metadata_make.R
\name{uas_metadata_make}
\alias{uas_metadata_make}
\title{Create a blank metadata text file}
\usage{
uas_metadata_make(
  dirs,
  file = "metadata.txt",
  suffix = c("", "dir")[1],
  flds = uas_getflds(),
  overwrite = FALSE,
  open = FALSE,
  quiet = FALSE
)
}
\arguments{
\item{dirs}{A character vector of directories or a uas_info object}

\item{file}{Name of the file to create}

\item{suffix}{A suffix to append to the file names}

\item{flds}{A character vector of field names}

\item{overwrite}{Overwrite existing metadata files, logical}

\item{open}{Open the file for editing}

\item{quiet}{Suppress messages}
}
\value{
A character vector of the external metadata text file(s) created.
}
\description{
Create a blank metadata text file
}
\details{
This creates blank external metadata text file(s). One file will be placed in each
directory in \code{dirs} (which can be either a character vector or a
\code{\link{uas_info}} object. The files will be named based on the value of \code{file}
which will be repeated if needed, with an option to append a suffix (which can make it
easier to edit them in a text editor). If \code{suffix = 'dir'}, the base name of the
directory will be appended. Suffix can also be a literal string. Both \code{file}
and \code{suffix} should be of length 1 or equal to the number of \code{dirs}.
}
\seealso{
\code{\link{uas_setflds}}
}
