% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uas_move.R
\name{uas_move}
\alias{uas_move}
\title{Move UAS images into sub-directories by group}
\usage{
uas_move(x, copymove = "move", basedir = NULL, quiet = FALSE)
}
\arguments{
\item{x}{A list of class 'uas_info'}

\item{copymove}{Move or copy}

\item{basedir}{Base dir}

\item{quiet}{Suppress messages}
}
\value{
A \link{uas_info} object
}
\description{
Move UAS images into sub-directories by group
}
\details{
NOT YET SUPPORTED

This will move groups of image files into subdirectories based on the group they fall into.
}
\seealso{
\link{uas_info}
}
