% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uas_getflds.R
\name{uas_setflds}
\alias{uas_setflds}
\title{Set default metadata fields}
\usage{
uas_setflds(flds = NULL, reset = FALSE)
}
\arguments{
\item{flds}{Name of metadata fields}

\item{reset}{Use orginal values}

\item{quiet}{Show messages, logical}
}
\description{
Set default metadata fields
}
