% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uas_worldfile.R
\name{uas_worldfile}
\alias{uas_worldfile}
\title{Create world and projection files for UAS images}
\usage{
uas_worldfile(
  x,
  idx = NULL,
  aux.xml = TRUE,
  wld = FALSE,
  wldext = "auto",
  prj = FALSE,
  rotated = TRUE,
  quiet = FALSE
)
}
\arguments{
\item{x}{A list of class 'uas_info'}

\item{idx}{Which rows in x to process (default is all)}

\item{aux.xml}{Create an aux.xml file, logical. See details.}

\item{wld}{Create a world file, logical. See details.}

\item{wldext}{Extension for the world file, character. Ignored if \code{wld = FALSE}.}

\item{prj}{Create a prj file, logical. See details.}

\item{rotated}{Compute parameters that replicate the camera Yaw.}

\item{quiet}{Show messages.}
}
\value{
A vector of filenames generated.
}
\description{
Create world files and projection files for UAS images
}
\details{
This function creates \href{https://en.wikipedia.org/wiki/Sidecar_file}{sidecard}
\href{https://en.wikipedia.org/wiki/World_file}{world files} and/or
projection files that are required to view raw images in GIS software such as
ArcGIS and QGIS.

\emph{Note: this function has been tested with JPG files from several DJI cameras. It has not yet been adapted for TIF files from multispectral cameras, and may  not work with those format.}

Note that the parameters in the world file are taken exclusively from the image
metadata, and will be \strong{approximate at best}. For a more accurate image placement,
process the images with photogrammetry software.

If your objective is to open the images in ArcGIS or QGIS, then generating
the \code{aux.xml} files should be all you need. \code{aux.xml} files
are \href{http://desktop.arcgis.com/en/arcmap/latest/manage-data/raster-and-images/auxiliary-files.htm}{ESRI auxillary files}
for raster layers. To be read by GIS software, they should have the same name as
the image file with 'aux.xml' added on. aux.xml files can contain a lot of info
about a raster layer, including statistics, the rotation info, coordinate
reference system, and other stuff. If generated by this function, they will
contain projection and rotation info only. Note also \strong{any existing aux.xml
files will be overwritten}. aux.xml files are the only sidecar file that ArcGIS
software supports for reading the projection info.

World files are small text files with extensions \code{jpw} and \code{tfw} for
JPG and TIF files respectively. To be read by GIS software, they must
have the same basename as the image and be saved in the same directory. Both
ArcGIS and QGIS read world files, however they are not needed if the same
info is available in an aux.xml file.

\code{prj} files contain just the Coordinate Reference System info. They do not
seem to be recognized for rasters by ArcGIS, however QGIS picks them up.
}
\seealso{
\link{uas_info}
}
