% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unit_conversions.R
\name{m2ft}
\alias{m2ft}
\alias{ft2m}
\alias{cm2in}
\alias{msq2acres}
\title{Unit conversion functions}
\usage{
m2ft(x)

ft2m(x)

cm2in(x)

msq2acres(x)
}
\arguments{
\item{x}{A numeric vector to convert}
}
\value{
A vector of converted values
}
\description{
Unit conversion functions
}
\details{
These unit conversion functions are vectorized.
}
\section{Functions}{
\itemize{
\item \code{ft2m()}: Feet to meters

\item \code{cm2in()}: Centimeters to inches

\item \code{msq2acres()}: Meters squared to acres

}}
