% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.uas_info.R
\name{print.uas_info}
\alias{print.uas_info}
\title{Print a summary of UAS metadata}
\usage{
\method{print}{uas_info}(x, metadata = TRUE, ...)
}
\arguments{
\item{x}{A UAS Image Collection object}

\item{metadata}{Show additional meta data info}

\item{...}{Other arguments (unused)}
}
\description{
Print a summary of metadata from a UAS image collection
}
\details{
Prints a summary of a drone images metadata collection object
}
\seealso{
\link{uas_info}
}
