% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uas_cameras.R
\name{uas_cameras}
\alias{uas_cameras}
\title{Cameras}
\usage{
uas_cameras()
}
\value{
A character vector containing the names of cameras known in the uavimg package
}
\description{
Returns the names of known cameras
}
\details{
In order to estimate the GSD and image footprint, certain properties of the camera are required, including the physical dimension of the sensor. Properties of several commonly used cameras have been collected and are stored in a csv file. If your camera is not listed, you may edit the csv file yourself (look in the directory where your R packages are saved), or contact the package author to have it added.
}
