% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uas_info.R
\name{uas_info}
\alias{uas_info}
\title{Extract info from UAS Images}
\usage{
uas_info(
  img_dirs,
  exiftool = NULL,
  csv = NULL,
  alt_agl = NULL,
  fp = TRUE,
  fwd_overlap = fp,
  cameras = NULL,
  metadata = "metadata.*\\\\.txt",
  cache = NULL,
  update_cache = FALSE,
  quiet = FALSE
)
}
\arguments{
\item{img_dirs}{Directory(s) where the image files reside}

\item{exiftool}{The path to the exiftool command line tool (omit if on the OS path)}

\item{csv}{The file name of a new csv file where the exif data will be saved (omit to make a temp one)}

\item{alt_agl}{The elevation above ground level in meters (optional for images with the relevative altitude saved)}

\item{fp}{Compute image foot prints, T/F}

\item{fwd_overlap}{Whether or not to compute the amount of overlap between one image and the next, T/F}

\item{cameras}{Location of the cameras.csv file. Is NULL the package csv file will be used.}

\item{metadata}{A filename pattern for a metadata file, or a metadata list object (see Details)}

\item{cache}{Logical or a directory where EXIF data should be cached (see Details)}

\item{update_cache}{Whether to update the cache}

\item{quiet}{Don't show messages}
}
\value{
A UAS Image Collection object. This is a named list with elements for
image centroids (as a sf data frame), footprints, total area, minimum convex polygon,
total directory size, the data flown, and external metadata.
}
\description{
Extracts info from geotagged images taken from a drone
}
\details{
This will read the EXIF header data from a directory of image files, and put out the centroids and image footprints on the ground. Mapping  image locations requires that the images have geostamps. In addition, mapping the image footprints requires that the camera parameters are known, and the flight elevation about ground level is either saved in the EXIF info, or provided in the \code{alt_agl} argument (in meters). If \code{alt_agl} is passed, it will override any elevation data in the EXIF info.

Camera parameters are saved in a csv file called cameras.csv. To see where this is saved, type \code{system.file("cameras/cameras.csv", package="uasimg")}. If your camera is not detected, please contact the package author or create an issue on \href{https://github.com/UCANR-IGIS/uasimg/issues}{GitHub}.

This function uses a free command line tool called EXIFtool to read the EXIF data, which can be downloaded
from \url{http://www.sno.phy.queensu.ca/~phil/exiftool/}.  After you download it, rename the executable file,
'exiftool(-k).exe' to 'exiftool.exe', and save it somewhere on your system's PATH (e.g., c:\\Windows).

\code{metadata} is an optional argument to pass external metadata that can not be extracted from the
images (e.g., location name, pilot). It can be a regular expression for a metadata filename
(in YAML format, see below) that should be in the same folder as the images. Or it can be a named list containing
metadata fields. For supported field names, see \code{\link{uas_getflds}}.

\code{metadata} cam also be a pattern for a filename (see \code{\link{list.files}}). If multiple files match the
pattern, they will all be read. Metadata files should be plain text in YAML format. Each line should contain a
key:value pair (with no quotes or delimiters). Lines starting with a hash or forward slash will be interpreted
as comments and ignored. For example:

\code{collection_name: Hopland Research and Extension Center, Watershed II
data_url: https://ucdavis.box.com/s/dp0sdfssxxxxxsdf
description: These data were collected as part of a pre-restoration monitoring program}

\code{cache} can be a logical value or the name of a directory where EXIF data gets cached.
If \code{cache = TRUE}, the default cache directory will be used (\code{~/.R}). Cached EXIF data is linked
to a directory of images based on the directory name and total size of all image files.
So if images are added or removed from the directory, the cache will be automatically rebuilt
the next time the function is run. \code{update_cache} is a logical value
which forces an update of cached data when TRUE.
}
\seealso{
\code{\link{uas_getcache}}, \code{\link{uas_report}}, \code{\link{uas_exp}}
}
