% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uas_info.R
\name{uas_info}
\alias{uas_info}
\title{Extract info from UAS Images}
\usage{
uas_info(
  img_dirs,
  exiftool = NULL,
  csv = NULL,
  alt_agl = NULL,
  fwd_overlap = TRUE,
  cameras = NULL,
  meta_extra = NULL,
  cache_dir = NULL,
  cache_update = FALSE,
  quiet = FALSE
)
}
\arguments{
\item{img_dirs}{Directory(s) where the image files reside}

\item{exiftool}{The path to the exiftool command line tool (omit if on the OS path)}

\item{csv}{The file name of a new csv file where the exif data will be saved (omit to make a temp one)}

\item{alt_agl}{The elevation above ground level in meters (optional for images with the relevative altitude saved)}

\item{fwd_overlap}{Whether or not to compute the amount of overlap between one image and the next, T/F}

\item{cameras}{Location of the cameras.csv file. Is NULL the package csv file will be used.}

\item{meta_extra}{A list of additional meta data (see Details)}

\item{cache_dir}{Directory to cache exif data}

\item{cache_update}{Whether to update the cached}

\item{quiet}{Don't show messages}
}
\value{
A named list with three elements: 1) SpatialPointsDataFrame with the image centroids, and 2) a SpatialPolygonsDataFrame of the image footprints, and 3) the image directory. An HTML report of the images can be created with \link{uas_report}.
}
\description{
Extracts info from geotagged images taken from a drone
}
\details{
This will read the EXIF header data from a directory of image files, and put out the centroids and image footprints on the ground. Mapping  image locations requires that the images have geostamps. In addition, mapping the image footprints requires that the camera parameters are known, and the flight elevation about ground level is either saved in the EXIF info, or provided in the \code{alt_agl} argument (in meters). If \code{alt_agl} is passed, it will override any elevation data in the EXIF info.

Camera parameters are saved in a csv file called cameras.csv. To see where this is saved, type \code{system.file("cameras/cameras.csv", package="uasimg")}. If your camera is not detected, please contact the package author or create an issue on \href{https://github.com/UCANR-IGIS/uasimg/issues}{GitHub}.

This function uses a free command line tool called EXIFtool to read the EXIF data, which can be downloaded
from \url{http://www.sno.phy.queensu.ca/~phil/exiftool/}.  After you download it, rename the executable file,
'exiftool(-k).exe' to 'exiftool.exe', and save it somewhere on your system's PATH (e.g., c:\\Windows).

\code{meta_extra} is an optional named list containing additional metadata field. Supported list elements include
\code{collection_name}, \code{data_url}, \code{description} and \code{contact}.

If \code{meta_extra} is not passed, an alternative way to pass these additional metadata
fields is to put them in a text file called 'meta_extra.txt' in the same directory as the images. The structure of the
text file should be a key:value pair on each line (no delimiters). For example:

collection_name: Hopland Research and Extension Center, Watershed II

Lines starting with a hash or forward slash will be interpreted as comments and ignored.

\code{cache_dir} is the name of a directory where EXIF data gets cached
}
\seealso{
\link{uas_report}, \link{uas_exp}
}
