% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uas_metadata_make.R
\name{uas_metadata_make}
\alias{uas_metadata_make}
\title{Manage supplemental metadata files}
\usage{
uas_metadata_make(
  x,
  md_file = "metadata.txt",
  md_suffix = NULL,
  make_new = FALSE,
  overwrite = FALSE,
  open = TRUE,
  md_flds = uas_getflds(),
  md_template = NULL,
  read_uasinfo = FALSE,
  use_system_editor = FALSE,
  quiet = FALSE
)
}
\arguments{
\item{x}{A character vector of directories, or a uas_info object}

\item{md_file}{Name of the file to create}

\item{md_suffix}{A suffix to append to the file names}

\item{make_new}{Create new metadata file(s), Logical}

\item{overwrite}{Overwrite existing metadata files, Logical}

\item{open}{Open the file for editing}

\item{md_flds}{A character vector of field names}

\item{md_template}{A template metadata file}

\item{read_uasinfo}{Read and populate metadata fields from x (assuming x is a uas_info object)}

\item{use_system_editor}{Whether to open metadata files in the operating system text editor (as opposed to RStudio), Logical}

\item{quiet}{Suppress messages}
}
\value{
A character vector of the external metadata text file(s) created.
}
\description{
Create and edit supplemental metadata files
}
\details{
This creates, and/or opens for editing, supplemental metadata text file(s) in image folders.

If \code{make_new = TRUE}, metadata files will be created in each directory of \code{x}. \code{x}
can be either a character vector or a \code{\link{uas_info}} object. The files will be named based on the value of \code{md_file},
with an option to append a suffix \code{md_suffix}. Both \code{md_file}
and \code{md_suffix} should be of length 1 or equal to the number of \code{x}.

If \code{open = TRUE}, the text files will be opened. To open all metadata text files,
}
\seealso{
\code{\link{uas_getflds}}, \code{\link{uas_setflds}}
}
