% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uas_metadata_make.R
\name{uas_metadata_make}
\alias{uas_metadata_make}
\title{Create flight metadata text files}
\usage{
uas_metadata_make(
  x,
  md_file = "metadata.txt",
  md_suffix = NULL,
  make_new = FALSE,
  overwrite = FALSE,
  open = TRUE,
  md_flds = uas_getflds(),
  md_template = NULL,
  read_uasinfo = FALSE,
  use_system_editor = FALSE,
  quiet = FALSE
)
}
\arguments{
\item{x}{A character vector of directories, or a uas_info object}

\item{md_file}{Name of the file to create}

\item{md_suffix}{A suffix to append to the file names}

\item{make_new}{Create new metadata file(s), Logical}

\item{overwrite}{Overwrite existing metadata files, Logical}

\item{open}{Open the file for editing}

\item{md_flds}{A character vector of field names}

\item{md_template}{A template metadata file}

\item{read_uasinfo}{Incorporate existing flight metadata already saved in x (assuming x is a uas_info object)}

\item{use_system_editor}{Whether to open metadata files in the operating system text editor (as opposed to RStudio), Logical}

\item{quiet}{Suppress messages}
}
\value{
A character vector of the external metadata text file(s) created.
}
\description{
Create and/or edit flight metadata text files
}
\details{
This creates, and/or opens for editing, supplemental metadata text file(s) in image folders.

If \code{make_new = TRUE}, metadata files will be created in each directory of \code{x}. \code{x}
can be either a character vector or a \code{\link{uas_info}} object. The files will be named based on the value of \code{md_file},
with an option to append a suffix \code{md_suffix}. Both \code{md_file}
and \code{md_suffix} should be of length 1 or equal to the number of \code{x}.

If \code{open = TRUE}, the text file(s) will be opened. To open existing metadata text files, let \code{open = TRUE}
and \code{make_new = FALSE}. By default, files will be opened with whichever text editor is in use by
R (see \code{file.edit}). To use the system text editor, let \code{use_system_editor = TRUE}.

To customize the fields that are added to flight metadata files, run \code{\link{uas_setflds}} first. To pre-populate some of the fields, pass
the path to an existing metadata file as the value for \code{md_template}. \code{md_template} can either be a local file or
online (\href{https://gist.githubusercontent.com/ajlyons/d0826f4775413ba27a21c62aff619bc2/raw/a6f3dde133652a729283b0e03f8a923f54f73a67/hrec_riverfire2020_metadata}{example}).
}
\seealso{
\code{\link{uas_getflds}}, \code{\link{uas_setflds}}
}
