% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uas_setflds.R
\name{uas_setflds}
\alias{uas_setflds}
\alias{uas_getflds}
\alias{uas_flds_oem}
\title{Manage supplemental metadata fields}
\usage{
uas_setflds(flds = NULL, reset = FALSE, quiet = FALSE)

uas_getflds()

uas_flds_oem()
}
\arguments{
\item{flds}{Character vector of metadata fields}

\item{reset}{Use a default set of metadata fields}

\item{quiet}{Suppress messages}
}
\description{
Set and view a default set of supplemental metadata fields
}
\details{
These functions allow you to define a default set of supplemental metadata fields that can be assigned
image collections either as a function argument, or a small text file placed
in the directory with the images. Note in this context 'supplemental metadata' refers to bits of information
that can not be automatically extracted from the images themselves (e.g., the name of the project
or pilot.)

Creating a default set of supplemental metadata fields does not make them mandatory, nor does
it limit the metadata fields you can use. The default set serves as a template when
a new metadata.txt file is created.
}
\section{Functions}{
\itemize{
\item \code{uas_setflds()}: Set default fields for image collection metadata.

\item \code{uas_getflds()}: Get default supplemental metadata field names.

\item \code{uas_flds_oem()}: Get a standard set of image collection metedata fields

}}
\seealso{
\code{\link{uas_getcache}}, \code{\link{uas_info}}
}
