% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uas_thumbnails_make.R
\name{uas_thumbnails_make}
\alias{uas_thumbnails_make}
\title{Create image thumbnails}
\usage{
uas_thumbnails_make(
  x,
  flt_idx = NULL,
  output_dir = NULL,
  tb_width = 400,
  overwrite = FALSE,
  use_magick = FALSE,
  stats = FALSE,
  quiet = FALSE
)
}
\arguments{
\item{x}{A list of class 'uas_info'}

\item{flt_idx}{Flight indices in x to process, integer}

\item{output_dir}{Output directory}

\item{tb_width}{Thumbnail width}

\item{overwrite}{Overwrite existing files}

\item{use_magick}{Use functions from the magick package}

\item{stats}{Report the amount of time it takes to create each thumbnail, logical}

\item{quiet}{Suppress messages}
}
\value{
A named list (one element for each directory processed) of thumbnail files created in the output directory
}
\description{
Move UAS images into sub-directories by group
}
\details{
This will create thumbnail images for the drone images in \code{x}. \code{flt_idx} allows you to specify a subset of flights
in \code{x} to process. The default output folder is a sub-directory of each image
folder called \emph{map/tb}, which will be created if needed. This location can be overridden with \code{output_dir}.
The dimensions of the thumbnails is determined by \code{tb_width}, from which the height is set automatically.

Thumbnail files will be given an 8-character suffix that looks random but is actually generated from the image contents.
This is to prevent clashes when thumnbail files from different flights are 'gathered' into a single folder attached to
a Table of Contents folder (see \code{\link{uas_toc}}).

If \code{use_magick = TRUE}, it will use resizing functions from the \link[magick]{magick} package. This is slower than the equivalent
functions from the \code{imager} package (the default), but may be necessary if you are processing TIFs and don't have
\href{https://imagemagick.org}{ImageMagick} installed on your computer (which \code{imager} requires to read TIFs).
}
\seealso{
\link{uas_report}
}
