% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uas_worldfile.R
\name{uas_worldfile}
\alias{uas_worldfile}
\title{Create world and projection files for UAS images}
\usage{
uas_worldfile(
  x,
  idx = NULL,
  aux.xml = TRUE,
  wld = FALSE,
  wldext = "auto",
  prj = FALSE,
  rotated = TRUE,
  quiet = FALSE
)
}
\arguments{
\item{x}{A list of class 'uas_info'}

\item{idx}{Which rows in x to process (default is all)}

\item{aux.xml}{Create an aux.xml file, logical. See details.}

\item{wld}{Create a world file, logical. See details.}

\item{wldext}{Extension for the world file, character. Ignored if \code{wld = FALSE}.}

\item{prj}{Create a prj file, logical. See details.}

\item{rotated}{Compute parameters that replicate the camera Yaw.}

\item{quiet}{Show messages.}
}
\value{
A vector of filenames generated.
}
\description{
Create world files and projection files for UAS images
}
\details{
This function creates \href{https://en.wikipedia.org/wiki/Sidecar_file}{sidecard}
\href{https://en.wikipedia.org/wiki/World_file}{world files} and/or
projection files that are required to view raw images in GIS software such as
ArcGIS and QGIS.

The parameters this function uses to generate world file are taken exclusively from the image
metadata, including the altitude above the launch point, the focal length, and so on. They should be considered
\strong{approximate at best}. For a more accurate image placement, process the images with photogrammetry software.


To generate world files, you must tell it to compute footprints when generating the flight metadata object. In other words, when you
run \code{\link{uas_info}}, pass \code{fp = TRUE}. If you've already generated a metadata object for the flight, you
may also need to pass \code{update_cache = TRUE}. Computing footprints requires an altitude for each image, which not
all drones support (especially multispectral cameras). For details see \code{\link{uas_info}}.

This function has been tested with JPG files from several DJI cameras. It has not yet been adapted for TIF files from
multispectral cameras, and may not work with those formats (contact the package author if you want to try).

 \code{aux.xml} files are
 \href{http://desktop.arcgis.com/en/arcmap/latest/manage-data/raster-and-images/auxiliary-files.htm}{ESRI auxillary files}
for raster layers.  If your objective is to open the images in ArcGIS or QGIS, then generating
the \code{aux.xml} files (the default setting) should be all you need. To be read by GIS software, they should
have the same name as the image file with 'aux.xml' added on. aux.xml files can contain a lot of info
about a raster layer, including statistics, the rotation info, coordinate
reference system, and other stuff. The files generated by this function will only contain projection
and rotation info only. Note also \strong{any existing aux.xml
files will be overwritten}. aux.xml files are the only sidecar file that ArcGIS
software seems to support for reading the projection info.

World files are small text files with extensions \code{jpw} and \code{tfw} for
JPG and TIF files respectively. To be read by GIS software, they must
have the same basename as the image and be saved in the same directory. Both
ArcGIS and QGIS read world files, however they are not needed if the same
info is available in an aux.xml file.

\code{prj} files contain just the Coordinate Reference System info. They do not
seem to be recognized for rasters by ArcGIS, however QGIS picks them up.
}
\seealso{
\code{\link{uas_info}}
}
